moved stuff into /src, created a /tests folder
