+++ src/database.js
@@ -1,7 +1,7 @@
 const { Sequelize } = require('sequelize');
 
 
-const sequelize = new Sequelize("bpa-db", "user", "pass", {
+const sequelize = new Sequelize("bpa-db", "root", "password", {
   dialect: "sqlite",
   host: "./bpa.sqlite",
   logging: false
