+++ src/encryption.js
